Singup & Dashboard
Updated the dashboard layout and changed the birthday field on the signup. The birthday now gets put into the database, as well as the UID which is the ID that firebase automatically gives users when they sign up. this will allow us to reference the user account which should help with deleting a profile